requirements.txt
flask
flask-cors
python-dotenv
google-genai
requests
gunicorn